build 1500

+++ source/io/github/xjrga/snack/other/CunninghamFormula.java
@@ -1,6 +1,7 @@
 package io.github.xjrga.snack.other;
 
 public class CunninghamFormula {
+
   private final Double weightinlbs;
 
   public CunninghamFormula(Double weightinlbs) {
@@ -8,6 +9,6 @@ public class CunninghamFormula {
   }
 
   public double getCalories() {
-    return 500.0D + (double) (22 * weightinlbs.intValue()) / 2.20462262185D;
+    return 500.0D + 22 * weightinlbs.intValue() / 2.20462262185D;
   }
 }
