+++ source/io/github/xjrga/snack/other/KatchMcArdleFormula.java
@@ -1,6 +1,7 @@
 package io.github.xjrga.snack.other;
 
 public class KatchMcArdleFormula {
+
   private final Double weightinlbs;
 
   public KatchMcArdleFormula(Double weightinlbs) {
@@ -8,6 +9,6 @@ public class KatchMcArdleFormula {
   }
 
   public double getCalories() {
-    return 370.0D + 21.6D * (double) weightinlbs.intValue() / 2.20462262185D;
+    return 370.0D + 21.6D * weightinlbs.intValue() / 2.20462262185D;
   }
 }
